Fixed bugs in the wifi_localization
The eigen library that was installed on this pc sometimes wasn't able to compute the correct inverse matrix. So a newer version is used now.
In the optimizer class in one case eta_minus was used where it should be eta_plus.
In the wifi_localization_node the probability values of each access point were summed, now they are multiplied.
The training_coordinates and obervations in the gaussian process are now get normalized for faster and better hyperparameter optimization.