Added nodes and launch files for experiments.
The map_traverser_node was rewritten. Now it simply takes a vector of coordinates and navigates to them.
The different nodes for the experiments publish needed data, so for example the difference between the position according to amcl and the position from the wifi position estimation.
Also there is a new message-type for the wifi position estimation. This is used when it is not supposed to publish an initial pose to amcl and one only needs the data published.